Add custom docrine types before registering otherwise we can't use custom types
