+++ src/EntityManagerFactory.php
@@ -9,7 +9,6 @@
 
 namespace ContainerInteropDoctrine;
 
-use Doctrine\DBAL\Types\Type;
 use Doctrine\ORM\EntityManager;
 use Interop\Container\ContainerInterface;
 
@@ -19,17 +18,10 @@ use Interop\Container\ContainerInterface;
 class EntityManagerFactory extends AbstractFactory
 {
     /**
-     * @var bool
-     */
-    private static $areTypesRegistered = false;
-
-    /**
      * {@inheritdoc}
      */
     protected function createWithConfig(ContainerInterface $container, $configKey)
     {
-        $this->registerTypes($container);
-
         $config = $this->retrieveConfig($container, $configKey, 'entity_manager');
 
         return EntityManager::create(
@@ -57,32 +49,5 @@ class EntityManagerFactory extends AbstractFactory
             'connection' => $configKey,
             'configuration' => $configKey,
         ];
-    }
-
-    /**
-     * Registers all declared typed, if not already done.
-     *
-     * @param ContainerInterface $container
-     */
-    private function registerTypes(ContainerInterface $container)
-    {
-        if (self::$areTypesRegistered) {
-            return;
-        }
-
-        $applicationConfig = $container->has('config') ? $container->get('config') : [];
-        $doctrineConfig = array_key_exists('doctrine', $applicationConfig) ? $applicationConfig['doctrine'] : [];
-        $typesConfig = array_key_exists('types', $doctrineConfig) ? $doctrineConfig['types'] : [];
-
-        foreach ($typesConfig as $name => $className) {
-            if (Type::hasType($name)) {
-                Type::overrideType($name, $className);
-                continue;
-            }
-
-            Type::addType($name, $className);
-        }
-
-        self::$areTypesRegistered = true;
     }
 }
